A little houskeeping.
Increasing counter sizes from int to long for averages.
Rename random functions to end with _co instead of _exclusive to
indicate range [close, open[, and to allow future suffixes oc, cc and oo.

+++ include/maxscale/response_stat.hh
@@ -36,7 +36,7 @@ public:
     void              query_ended();    // ok to call without a query_started
     bool              make_valid();     // make valid even if there are only filter_samples
     bool              is_valid() const;
-    int               num_samples() const;
+    long              num_samples() const;
     maxbase::Duration average() const;
     bool              sync_time_reached();  // is it time to apply the average to the server?
     void              reset();
@@ -44,7 +44,7 @@ public:
 private:
     const int                      m_num_filter_samples;
     const maxbase::Duration        m_sync_duration;
-    int                            m_sample_count;
+    long                           m_sample_count;
     std::vector<maxbase::Duration> m_samples;   // N sampels from which median is used
     maxbase::CumulativeAverage     m_average;
     maxbase::TimePoint             m_last_start;
